+++ CHANGELOG.md
@@ -4,6 +4,7 @@ Changelog
 ## [Unreleased]
 
 - Slow operations are prohibited on EDT
+- Can't add entry to ignore file via context menu
 
 ## [v4.1.1] (https://github.com/JetBrains/idea-gitignore/tree/v4.1.1) (2021-06-02)