+++ CHANGELOG.md
@@ -2,6 +2,7 @@ Changelog
 ----------
 
 ## [Unreleased]
+## [vextension 'changelog' property 'version'] (https://github.com/JetBrains/idea-gitignore/tree/vextension 'changelog' property 'version') (2021-07-28)
 
 - Removed "Hide ignored files" feature in favor of [Foldable ProjectView](https://github.com/hsz/intellij-foldable-projectview) plugin
 - Fixed: Slow operations are prohibited on EDT